Add Support for Hyperledger Besu
Most of the existing ethereum adapter works fine for Besu except for two
major issues:  Besu does not have wallet support (and never will as an
architectual choice) and hence the automatic management of nonces no
longer works.

First, we add two new config vars contractDeployerAddressPrivateKey and
fromAddressPrivateKey and if present prefer the 'raw' transaciton APIs.

Second in the main transaction loop we track the nonce of the current
transaction so this can set the nonces appropriately.

Finally we add in gas estimation rather than a hard coded gas number
so that blocks can be as full as possible.

Besu has a slightly different genesis config from geth, so a besu config
directory in the samples is added. The samples network only uses Clique
right now because PoW is not something enterprises should be deploying.

Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>