Remove redundant event code from v1 gateway
For a gateway implementation it cannot support compatibility mode.

The gateway implementation should only be used against a fabric 1.4
network and a 1.4 peer does not have the old eventhub capability
anymore. The node-sdk also removed the old eventhub implementation in
fabric-client.

The other compatibility checks are also redundant now

This PR removes the redundant code.

Signed-off-by: D Kelsey <d_kelsey@uk.ibm.com>