Fixes caliper hang of unfinished transactions
This addresses a long standing issue in caliper where a round will hang
waiting for unfinished transactions that have actually finished but not
been recorded.

It also addresses an issue where the fabric connectors try to change the
time_create value of a TxStatus but in fact actually doesn't change the
creation time.

closes #1068
closes #1340

Signed-off-by: D <d_kelsey@uk.ibm.com>