Remove fisco-bcos (#1515)
Node 16 is now out of lts. Fisco bcos will not work on node 18 or higher
due to it not being compatible with openssl 3 so caliper at the moment
can't move it's builds to node 18 or node 20. Also openssl 1.1 is now
out of support and so the connector is using crypto libraries which are
no longer supported. As we have no-one who can maintain this connector,
it should be dropped.

Signed-off-by: Dave Kelsey <d_kelsey@uk.ibm.com>